refactor(globe): modularize WebGL globe implementation
Reorganizes the WebGL globe composable into a modular architecture by splitting functionality into separate, focused modules for better maintainability and testability.

Extracts color parsing with LRU caching to optimize repeated color conversions and avoid expensive canvas operations.

Separates geometry generation (octahedron sphere tessellation and great-circle arcs) into dedicated utilities with improved ribbon rendering for visible arc width.

Isolates interaction handling (mouse, touch, inertia) into a standalone module with cleaner event management.

Moves shader definitions and texture generation logic into separate files for better organization.

Consolidates exports through a central index file, providing a clean public API while keeping internal utilities accessible for advanced usage.

Fixes coordinate handling by validating latitude/longitude values with finite number checks and defaulting to zero for invalid inputs, preventing potential NaN issues in globe positioning.

Improves arc animation with dashed rendering effect and triangle strip geometry instead of line strips for better visual quality.

Reduces near-trajectory arc spam by showing only destination ripples when source and target are within 5 degrees, with configurable delay timing.

Removes unused D3 dependencies (color, drag, ease, selection, timer, transition, zoom) that were replaced by custom implementations, reducing bundle size.

Adds nextTick await before showing globe to ensure proper DOM mounting timing.
