+++ package.json
@@ -43,16 +43,8 @@
     "@tailwindcss/vite": "^4.1.18",
     "@tanstack/vue-form": "^1.28.0",
     "@tanstack/vue-table": "^8.21.3",
-    "@types/d3-color": "^3.1.3",
-    "@types/d3-drag": "^3.0.7",
-    "@types/d3-ease": "^3.0.2",
     "@types/d3-geo": "^3.1.0",
     "@types/d3-scale": "^4.0.9",
-    "@types/d3-scale-chromatic": "^3.1.0",
-    "@types/d3-selection": "^3.0.11",
-    "@types/d3-timer": "^3.0.2",
-    "@types/d3-transition": "^3.0.9",
-    "@types/d3-zoom": "^3.0.8",
     "@types/node": "^25.2.0",
     "@unovis/ts": "^1.6.4",
     "@unovis/vue": "^1.6.4",
@@ -62,16 +54,8 @@
     "baseline-browser-mapping": "^2.9.19",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "d3-color": "^3.1.0",
-    "d3-drag": "^3.0.0",
-    "d3-ease": "^3.0.1",
     "d3-geo": "^3.1.1",
     "d3-scale": "^4.0.2",
-    "d3-scale-chromatic": "^3.1.0",
-    "d3-selection": "^3.0.0",
-    "d3-timer": "^3.0.1",
-    "d3-transition": "^3.0.1",
-    "d3-zoom": "^3.0.0",
     "destr": "^2.0.5",
     "embla-carousel-vue": "^8.6.0",
     "es-toolkit": "^1.44.0",
